Refactor: Separate frontend and backend concerns
- Created a 'backend' directory.
- Moved Firebase configuration, services, and mockData into the 'backend' directory.
- Updated import paths in frontend components to reflect the new backend file locations.
- Verified that backend import paths remain correct due to relative path usage.
