+++ app/chat/index.tsx
@@ -7,7 +7,7 @@ import { LinearGradient } from 'expo-linear-gradient';
 
 import ChatListItem from '@/components/chat/ChatListItem'; // Adjusted path
 import { ChatRoom, UserProfile } from '@/types'; // Adjusted path
-import { subscribeToUserChatRooms } from '@/services/firebaseService'; // Adjusted path
+import { subscribeToUserChatRooms } from '@/backend/services/firebaseService'; // Adjusted path
 import { useAuth } from '@/context/AuthContext'; // Import useAuth
 
 // Mock current user - Will be removed
