+++ backend/services/firebaseService.ts
